changes to do with gcc 4.5 to 4.6 change
It seems that the Numerical Recipes style of helper routines only visible from
within a given subroutine breaks in combination with namespaces in gcc4.6. I
now have a declaration and definition outside of the function within the same
namespace as the function. Not terribly satisfactory, but it seems to work.